requirements.txt
# Core
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20

# Vector Database
chromadb>=0.4.22

# Document Parsing
python-docx>=1.1.0
python-pptx>=0.6.23
pdfplumber>=0.10.0
PyMuPDF>=1.23.0

# AI-Powered Document Parser (Optional but Recommended)
docling>=2.0.0  # 支持 PDF、DOCX、PPTX 统一解析

# OCR - PaddleOCR (Mac 本地离线 OCR，支持 GPU 加速)
paddlepaddle  # Mac 原生支持，M 系列自动调用 Metal GPU
paddleocr     # PP-OCRv5 轻量级 OCR 模型

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
tqdm>=4.66.0

# Dev
pytest>=7.4.0
pytest-asyncio>=0.21.0